+++ qobuz-player-web/src/routes/album.rs
@@ -24,7 +24,7 @@ pub(crate) fn routes() -> Router<std::sync::Arc<crate::AppState>> {
 
 async fn play_track(
     State(state): State<Arc<AppState>>,
-    Path((id, track_position)): Path<(String, u32)>,
+    Path((id, track_position)): Path<(String, usize)>,
 ) -> impl IntoResponse {
     state.controls.play_album(&id, track_position);
 }
